requirements.txt
django-bootstrap-datepicker-plus
django-bootstrap4
channels
python-decouple
whitenoise
django-sslserver
av
pysubs2
tables
keyboard
pandas
scipy
h5py
pyyaml
matplotlib